# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tunnex API
  version: 0.1.0
  description: |
    Control-plane API for Tunnex.io. This spec is the single source of truth:
    the Go server types/handlers and the TypeScript client are generated from it
    (`make generate`), and drift fails the build (`make generate-check`).
servers:
  - url: /
    description: Same-origin (served behind the edge nginx)

# Endpoints require a session cookie OR a CLI bearer credential by default
# (bearer ≡ cookie, S5.1); public ones set `security: []`; cookie-only ops
# (credential minting) override with `security: [{cookieAuth: []}]` + argument.
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/auth/password:
    post:
      operationId: changePassword
      summary: Change your own password
      description: >-
        ⛔ The ONLY way out of a forced password change, and therefore the one authenticated route that is
        NOT blocked by one. The CP admin's bootstrap credential is printed to the logs, so it is treated as
        compromised the moment it works — without this endpoint the forced change would be a lockout.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 12 }
      responses:
        "204":
          description: Changed.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/auth/cli/authorize:
    post:
      operationId: cliAuthorize
      summary: Mint a one-time CLI authorization code (browser leg)
      description: >
        COOKIE-SESSION ONLY (argued exception to bearer≡cookie): minting a new
        credential from an existing bearer credential would be self-replication —
        a stolen token could outlive its expiry by re-minting. The browser session
        is the human checkpoint. Verified email required. The redirect_uri must be
        a loopback (http://127.0.0.1:<port>/callback or http://[::1]:<port>/callback,
        any port, exactly that path — never a hostname). The code is single-use,
        expires in 60s, and is bound to this exact redirect_uri + PKCE challenge.
      tags: [cli-auth]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CliAuthorizeRequest" }
      responses:
        "200":
          description: One-time code minted; the SPA redirects the browser to the loopback.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CliAuthorizeResponse" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/auth/cli/device/approve:
    post:
      operationId: cliDeviceApprove
      summary: Approve a device-code login (browser leg)
      description: >
        COOKIE-SESSION ONLY (same argued exception as cliAuthorize) + verified
        email required — this is the human checkpoint of the device flow.
      tags: [cli-auth]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CliDeviceApproveRequest" }
      responses:
        "200":
          description: Approved — the polling CLI will receive its credential.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenericMessage" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/gateway-endpoint:
    get:
      operationId: getGatewayEndpoint
      summary: Get the deployment-wide gateway control endpoint
      description: >-
        Returns the raw mTLS endpoint used by newly generated gateway join commands. This is a
        deployment setting, not an organization setting. Reading and changing it requires a verified
        deployment administrator or an owner/admin of an organization on this deployment.
      tags: [admin]
      responses:
        "200":
          description: Gateway control endpoint.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayEndpoint" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: updateGatewayEndpoint
      summary: Set the deployment-wide gateway control endpoint
      description: >-
        Sets the hostname and optional port used in future gateway join commands. The endpoint must be
        an absolute HTTPS URL with no path, query or fragment; it must terminate in the raw mTLS agent
        listener, normally port 8443. Existing gateways are not changed by this setting.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateGatewayEndpointRequest" }
      responses:
        "200":
          description: Updated gateway control endpoint.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayEndpoint" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/organizations:
    get:
      operationId: listOrganizations
      summary: List organizations
      description: Lists the organizations the authenticated user belongs to.
      tags: [organizations]
      responses:
        "200":
          description: Organizations.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createOrganization
      summary: Create an organization
      description: |
        Creates an organization. In the open edition this fails with
        `org_limit_reached` once one organization exists; the enterprise build
        allows multiple.
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: The created organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/organizations/{orgId}:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getOrganization
      summary: Get an organization
      tags: [organizations]
      responses:
        "200":
          description: The organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateOrganization
      summary: Update organization settings (name only; slug is immutable)
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: The updated organization.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteOrganization
      summary: Soft-delete an organization (refused while it still owns resources)
      description: >-
        Owner-only (`org:delete`). Refuses with 409 `org_not_empty` while the organization still has
        gateways, sites, devices, Kubernetes clusters or machine credentials — a soft delete leaves them
        running with no organization to manage them from. There is no force flag.
      tags: [organizations]
      responses:
        "204":
          description: Deleted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/organizations/{orgId}/deletion-preflight:
    parameters:
      - name: orgId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: orgDeletionPreflight
      summary: What still blocks deleting this organization
      description: >-
        The counts the delete refusal is derived from, read BEFORE anyone types a confirmation. Same
        permission as the delete itself — a preflight is only actionable to someone who could act on it.
      tags: [organizations]
      responses:
        "200":
          description: Preflight.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrgDeletionPreflight" }
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tunnex_session
    # CLI credential (tnx_...). bearer ≡ cookie on every authenticated endpoint;
    # the ONLY exceptions are argued on the ops themselves (cliAuthorize +
    # cliDeviceApprove — credential minting must not be reachable from a bearer).
    bearerAuth:
      type: http
      scheme: bearer
    runtimeBearer:
      type: http
      scheme: bearer
      description: Machine-only hash-backed managed-agent runtime credential; never a user/session bearer.
  headers:
    RequestId:
      description: Correlation ID for this request; also appears in server logs.
      schema:
        type: string
      required: true
  schemas:
...
    Organization:
      type: object
      required: [id, name, slug, pool_cidr, max_agent_identities, managed_agent_runtime_enabled, created_at, updated_at]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        pool_cidr:
          type: string
          description: The org's flat WireGuard address pool (IPv4 CIDR).
        ovpn_enabled:
          type: boolean
          description: "S9.1 D-S9.5-OPTIN: whether this org has opted into OpenVPN. Default false. When false the OpenVPN device type is not offered and the export endpoint refuses (opt_in_required)."
        max_agent_identities:
          type: integer
          format: int32
          minimum: 0
          nullable: true
          description: "F02: organization-wide managed-agent identity quota. null means unlimited; no remaining count is exposed."
        managed_agent_runtime_enabled:
          type: boolean
          description: "F04: explicit organization opt-in for managed runtime synchronization. Default false; a paid licence does not enable it implicitly."
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateOrganizationRequest:
      type: object
      required: [name, slug]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 63
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
    UpdateOrganizationRequest:
      type: object
      required: [name]
      additionalProperties: false
      description: Only the name is mutable; the slug is fixed at creation.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
...
    GatewayEndpoint:
      type: object
      required: [url, configured]
      additionalProperties: false
      properties:
        url:
          type: string
          description: The raw mTLS gateway control URL used by future join commands.
        configured:
          type: boolean
          description: False when the deployment has no explicit gateway endpoint yet.
    UpdateGatewayEndpointRequest:
      type: object
      required: [url]
      additionalProperties: false
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 300
          description: Absolute https URL for the raw mTLS agent listener, normally ending in :8443.

    # ---- CLI credential flow (S5.1) ----------------------------------------
    CliAuthorizeRequest:
      type: object
      required: [redirect_uri, code_challenge, state]
      additionalProperties: false
      properties:
        redirect_uri:
          type: string
          maxLength: 200
          description: >
            Loopback only — http://127.0.0.1:<port>/callback or
            http://[::1]:<port>/callback (any port, exactly that path; hostnames
            incl. "localhost" are refused — DNS-spoofable).
        code_challenge:
          type: string
          minLength: 43
          maxLength: 128
          description: PKCE S256 challenge (base64url SHA-256 of the CLI's verifier).
        state:
          type: string
          minLength: 1
          maxLength: 200
          description: Echoed to the loopback callback verbatim (CLI-side correlation/CSRF).
    CliAuthorizeResponse:
      type: object
      x-go-name: CliAuthorizeResult
      required: [code, state, expires_in]
      additionalProperties: false
      properties:
        code: { type: string, description: "One-time authorization code (single-use, hashed at rest)." }
        state: { type: string }
        expires_in: { type: integer, description: "Code lifetime in seconds (60)." }
...
    CliDeviceApproveRequest:
      type: object
      required: [user_code]
      additionalProperties: false
      properties:
        user_code: { type: string, minLength: 1, maxLength: 32 }
...
    GenericMessage:
      type: object
      required: [message]
      additionalProperties: false
      properties:
        message: { type: string }
...
    OrgDeletionPreflight:
      type: object
      additionalProperties: false
      required: [deletable, gateways, devices, sites, clusters, machine_credentials, blockers]
      properties:
        deletable:
          type: boolean
          description: True only when every count below is zero.
        gateways: { type: integer }
        devices: { type: integer }
        sites: { type: integer }
        clusters: { type: integer }
        machine_credentials: { type: integer }
        blockers:
          type: array
          description: >-
            Every blocker, already worded and pluralised for a human ("2 gateways", "1 site"). Served as
            text so the screen and the refusal message cannot drift into describing the same state
            differently.
          items: { type: string }
...
  responses:
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...